# Add stellar/go-xdr#16 (reduction on XDR encoding/decoding allocations)