--- a/go.list
+++ b/go.list
@@ -70,7 +70,7 @@ github.com/spf13/jwalterweatherman v0.0.0-20141219030609-3d60171a6431
 github.com/spf13/pflag v0.0.0-20161005214240-4bd69631f475
 github.com/spf13/viper v0.0.0-20150621231900-db7ff930a189
 github.com/stellar/go 
-github.com/stellar/go-xdr v0.0.0-20201028102745-f80a23dac78a
+github.com/stellar/go-xdr v0.0.0-20211103144802-8017fc4bdfee
 github.com/stellar/throttled v2.2.3-0.20190823235211-89d75816f59d+incompatible
 github.com/stretchr/objx v0.3.0
 github.com/stretchr/testify v1.7.0