This PR incorporates stellar/go-xdr#16  in Horizon.

The benchmarks (copied from https://github.com/stellar/xdrgen/pull/65 but using a larger transaction) show a big improvement in the allocations made by the encoder/decoder.


Before:

```                                                                                                                                                                                                                                                             
goos: darwin
goarch: amd64
pkg: github.com/stellar/go/benchmarks
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkXDRUnmarshal
BenchmarkXDRUnmarshal-8              100             18301 ns/op            5968 B/op        232 allocs/op
BenchmarkGXDRUnmarshal
BenchmarkGXDRUnmarshal-8             100             58419 ns/op           86738 B/op        278 allocs/op
BenchmarkXDRMarshal
BenchmarkXDRMarshal-8                100             13599 ns/op            5248 B/op        194 allocs/op
BenchmarkGXDRMarshal
BenchmarkGXDRMarshal-8               100              8561 ns/op            2152 B/op        157 allocs/op
PASS
```

After:

```
goos: darwin
goarch: amd64
pkg: github.com/stellar/go/benchmarks
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkXDRUnmarshal
BenchmarkXDRUnmarshal-8              100             18851 ns/op            5576 B/op        186 allocs/op
BenchmarkGXDRUnmarshal
BenchmarkGXDRUnmarshal-8             100             51694 ns/op           86736 B/op        278 allocs/op
BenchmarkXDRMarshal
BenchmarkXDRMarshal-8                100             11203 ns/op            5128 B/op        156 allocs/op
BenchmarkGXDRMarshal
BenchmarkGXDRMarshal-8               100              8539 ns/op            2152 B/op        157 allocs/op
PASS
```

Note how the allocations go down:
* from 232 to 186 (~20% reduction) for the decoder
* from 194 to 156 (also ~20% reduction) for the encoder


This should relieve a lot of pressure in the garbage collector ( https://github.com/stellar/go/issues/4022 shows how allocations/garbage collection constitute a big part of Horizon's runtime)